requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
Django==3.0.8
django-bootstrap4==2.2.0
django-chartjs==2.1.0
django-crispy-forms==1.9.2
django-money==1.1
Faker==4.1.1
importlib-metadata==1.7.0
mysqlclient==2.0.1
py-moneyed==0.8.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
text-unidecode==1.3
zipp==3.1.0
